Solved bugs 1-4

+++ mpi_solved2.c
@@ -1,8 +1,8 @@
 /******************************************************************************
 * FILE: mpi_bug2.c
 * DESCRIPTION: 
-*   This program has a bug that causes wrong answers and/or termination - depends
-*   upon the MPI library and platform.
+*   This program has a bug that causes wrong answers and/or termination - 
+*   depends upon the MPI library and platform.
 * SOURCE: Blaise Barney 
 * LAST REVISED: 01/24/09
 ******************************************************************************/
@@ -13,7 +13,9 @@
 int main (int argc, char *argv[])
 {
 int numtasks, rank, tag=1, alpha, i;
-float beta;
+// Task 0 sends ints, but task 1 expects floats.  The issue can be resolved by
+// choosing a common datatype (in this case, int).
+int beta;
 MPI_Request reqs[10];
 MPI_Status stats[10];
 
@@ -34,9 +36,9 @@ if (rank == 0) {
 
 if (rank == 1) {
   for (i=0; i<10; i++) {
-    MPI_Irecv(&beta, 1, MPI_FLOAT, 0, tag, MPI_COMM_WORLD, &reqs[i]);
+    MPI_Irecv(&beta, 1, MPI_INT, 0, tag, MPI_COMM_WORLD, &reqs[i]);
     MPI_Wait(&reqs[i], &stats[i]);
-    printf("Task %d received = %f\n",rank,beta);
+    printf("Task %d received = %d\n",rank,beta);
     }
   }
